Struct nautilus_accounting::position::Position

#[repr(C)]
pub struct Position {

    pub trader_id: TraderId,
    pub strategy_id: StrategyId,
    pub instrument_id: InstrumentId,
    pub id: PositionId,
    pub account_id: AccountId,
    pub opening_order_id: ClientOrderId,
    pub closing_order_id: Option<ClientOrderId>,
    pub entry: OrderSide,
    pub side: PositionSide,
    pub signed_qty: f64,
    pub quantity: Quantity,
    pub peak_qty: Quantity,
    pub price_precision: u8,
    pub size_precision: u8,
    pub multiplier: Quantity,
    pub is_inverse: bool,
    pub base_currency: Option<Currency>,
    pub quote_currency: Currency,
    pub settlement_currency: Currency,
    pub ts_init: UnixNanos,
    pub ts_opened: UnixNanos,
    pub ts_last: UnixNanos,
    pub ts_closed: Option<UnixNanos>,
    pub duration_ns: u64,
    pub avg_px_open: f64,
    pub avg_px_close: Option<f64>,
    pub realized_return: Option<f64>,
    pub realized_pnl: Option<Money>,
    /* private fields */

}

Represents a position in a financial market.

The position ID may be assigned at the trading venue, or can be system generated depending on a strategies OMS (Order Management System) settings.

Fields
trader_id: TraderIdstrategy_id: StrategyIdinstrument_id: InstrumentIdid: PositionIdaccount_id: AccountIdopening_order_id: ClientOrderIdclosing_order_id: Option<ClientOrderId>entry: OrderSideside: PositionSidesigned_qty: f64quantity: Quantitypeak_qty: Quantityprice_precision: u8size_precision: u8multiplier: Quantityis_inverse: boolbase_currency: Option<Currency>quote_currency: Currencysettlement_currency: Currencyts_init: UnixNanosts_opened: UnixNanosts_last: UnixNanosts_closed: Option<UnixNanos>duration_ns: u64avg_px_open: f64avg_px_close: Option<f64>realized_return: Option<f64>realized_pnl: Option<Money>
Implementations

impl Position

pub fn new<T: Instrument>(instrument: T, fill: OrderFilled) -> Result<Self>